+++ CHANGELOG.md
@@ -3,6 +3,7 @@ CHANGELOG
 
 5.0.0-beta5.pre
 ----------------
+ * Content elements redirrect is fixed
  * Cms user personal area
  
 5.0.0-beta4